refactor: module 이름 변경 (ecommerce service > product service)
