--- module-ecommerce-service/src/main/java/com/shop/ecommerceservice/EcommerceServiceApplication.java
@@ -1,16 +1,0 @@
-package com.shop.ecommerceservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
-public class EcommerceServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(EcommerceServiceApplication.class, args);
-    }
-
-}
