--- module-ecommerce-service/src/main/java/com/shop/ecommerceservice/common/exception/BaseException.java
@@ -1,19 +1,0 @@
-package com.shop.ecommerceservice.common.exception;
-
-import com.shop.ecommerceservice.common.response.ErrorCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class BaseException extends RuntimeException {
-
-    private ErrorCode errorCode;
-    private String message;
-
-    public BaseException(ErrorCode errorCode) {
-        this.errorCode = errorCode;
-        this.message = errorCode.getMessage();
-    }
-
-}
