--- module-ecommerce-service/src/main/java/com/shop/ecommerceservice/product/repository/ProductRepository.java
@@ -1,7 +1,0 @@
-package com.shop.ecommerceservice.product.repository;
-
-import com.shop.ecommerceservice.product.entity.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
