+++ module-product-service/build.gradle
