+++ module-product-service/src/main/java/com/shop/productservice/common/entity/BaseTimeEntity.java
@@ -1,0 +1,26 @@
+package com.shop.productservice.common.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+@Getter
+@EntityListeners(value = {AuditingEntityListener.class})
+@MappedSuperclass
+public abstract class BaseTimeEntity {
+
+    @CreatedDate
+    @Column(name = "created_at", nullable = false, updatable = false)
+    protected LocalDateTime createdAt; // 생성일시
+
+    @LastModifiedDate
+    @Column(name = "updated_at", nullable = false)
+    protected LocalDateTime updatedAt; // 수정일시
+
+}
