+++ module-product-service/src/main/java/com/shop/productservice/common/exception/BaseException.java
@@ -1,0 +1,19 @@
+package com.shop.productservice.common.exception;
+
+import com.shop.productservice.common.response.ErrorCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class BaseException extends RuntimeException {
+
+    private ErrorCode errorCode;
+    private String message;
+
+    public BaseException(ErrorCode errorCode) {
+        this.errorCode = errorCode;
+        this.message = errorCode.getMessage();
+    }
+
+}
