+++ module-product-service/src/main/java/com/shop/productservice/repository/ProductRepository.java
@@ -1,0 +1,7 @@
+package com.shop.productservice.repository;
+
+import com.shop.productservice.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+}
