+++ settings.gradle
@@ -4,4 +4,5 @@ include 'module-user-service'
 include 'module-activity-service'
 include 'module-newsfeed-service'
 include 'module-ecommerce-service'
+include 'module-product-service'
